Fixed: packageinfo was not updated in bundle. Bundle-Version was not set in the .bnd file

+++ bndtools.release/src/bndtools/release/ReleaseHelper.java
@@ -11,26 +11,30 @@
 package bndtools.release;
 
 import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.PrintWriter;
 import java.util.Collection;
 import java.util.List;
 
+import org.eclipse.core.resources.IFile;
 import org.eclipse.core.resources.IProject;
 import org.eclipse.core.resources.IResource;
+import org.eclipse.core.runtime.CoreException;
+import org.eclipse.jface.text.IDocument;
 import org.eclipse.swt.widgets.Display;
 import org.eclipse.swt.widgets.Shell;
 import org.eclipse.ui.PlatformUI;
 
 import aQute.bnd.build.Project;
+import aQute.bnd.service.RepositoryPlugin.Strategy;
+import aQute.lib.io.IO;
 import aQute.lib.osgi.Builder;
 import aQute.lib.osgi.Jar;
 import aQute.libg.reporter.Reporter;
-import aQute.libg.sed.Replacer;
-import aQute.libg.sed.Sed;
 import bndtools.diff.JarDiff;
 import bndtools.diff.PackageInfo;
+import bndtools.editor.model.BndEditModel;
 import bndtools.release.api.IReleaseParticipant;
 import bndtools.release.api.IReleaseParticipant.Scope;
 import bndtools.release.api.ReleaseContext;
@@ -42,42 +46,78 @@ public class ReleaseHelper {
 	public static void updateProject(ReleaseContext context) throws Exception {
 		
 		Collection<? extends Builder> builders = context.getProject().getBuilder(null).getSubBuilders();
-		for (Builder b : builders) {
-			JarDiff current = null;
-			for (JarDiff jd : context.getJarDiffs()) {
-				if (jd.getSymbolicName().equals(b.getBsn())) {
-					current = jd;
-					break;
-				}
+		for (Builder builder : builders) {
+
+			JarDiff current = getJarDiffForBuilder(builder, context);
+			if (current == null) {
+				continue;
 			}
-			if (current != null) {
-				String bundleVersion = current.getSuggestedVersion();
-				for (PackageInfo pi : current.getModifiedExportedPackages()) {
-					if (pi.getVersion() != null && !pi.getVersion().equals(pi.getSuggestedVersion())) {
-						updatePackageInfoFile(context.getProject(), pi);
-					}
-				}
-				
-				for (PackageInfo pi : current.getNewExportedPackages()) {
+			for (PackageInfo pi : current.getModifiedExportedPackages()) {
+				if (pi.getVersion() != null && !pi.getVersion().equals(pi.getSuggestedVersion())) {
 					updatePackageInfoFile(context.getProject(), pi);
 				}
-				
-				if (bundleVersion != null) {
-					File file = context.getProject().getPropertiesFile();
-					Replacer replacer = context.getProject().getReplacer();
-					if (b.getPropertiesFile() != null) {
-						file = b.getPropertiesFile();
-						replacer = b.getReplacer();
-					}
-					Sed sed = new Sed(replacer, file);
-					sed.replace("(Bundle-Version\\s*(:|=)\\s*)(([0-9]+(\\.[0-9]+(\\.[0-9]+)?)?))",
-		                    "$1${version;===;" + bundleVersion + "}");
-					sed.doIt();
-				}
 			}
-		}
+			
+			for (PackageInfo pi : current.getNewExportedPackages()) {
+				updatePackageInfoFile(context.getProject(), pi);
+			}
+
+			updateBundleVersion(context, current);
+   		}
 	}
 	
+	private static void updateBundleVersion(ReleaseContext context, JarDiff current) throws IOException, CoreException {
+		
+		String bundleVersion = current.getSuggestedVersion();
+		if (bundleVersion != null) {
+			
+			File file = context.getProject().getPropertiesFile();
+		
+			final IFile resource = (IFile) ReleaseUtils.toResource(file);
+			
+			IDocument document = FileUtils.readFully(resource);
+
+			final BndEditModel model = new BndEditModel();
+			model.loadFrom(document);
+
+			String savedVersion = model.getBundleVersionString();
+			if (savedVersion != null && savedVersion.indexOf('$') > -1) {
+				//TODO: Handle macros / variables
+			}
+			model.setBundleVersion(bundleVersion);
+			
+			final IDocument finalDoc = document;
+			Runnable run = new Runnable() {
+
+				public void run() {
+					model.saveChangesTo(finalDoc);
+					
+					try {
+						FileUtils.writeFully(finalDoc, resource, false);
+						resource.refreshLocal(IResource.DEPTH_ZERO, null);
+					} catch (CoreException e) {
+						throw new RuntimeException(e);
+					}
+				}
+			};
+	        if (Display.getCurrent() == null) {
+	            Display.getDefault().syncExec(run);
+	        } else
+	            run.run();
+		}
+	}
+
+	private static JarDiff getJarDiffForBuilder(Builder builder, ReleaseContext context) {
+		JarDiff current = null;
+		for (JarDiff jd : context.getJarDiffs()) {
+			if (jd.getSymbolicName().equals(builder.getBsn())) {
+				current = jd;
+				break;
+			}
+		}
+		return current;
+	}
+
 	public static boolean release(ReleaseContext context, List<JarDiff> jarDiffs) throws Exception {
 
 		boolean ret = true;
@@ -133,9 +173,13 @@ public class ReleaseHelper {
 		}
 	}
 	
+	private static void handleReleaseErrors(ReleaseContext context, Reporter reporter, String symbolicName, String version) {
+		for (String message : reporter.getErrors()) {
+			context.getErrorHandler().error(symbolicName, version, message);
+		}
+	}
+
 	private static void displayErrors(ReleaseContext context, Scope scope) {
-		
-		
 		
 		final String name = context.getProject().getName();
 		final List<Error> errors = context.getErrorHandler().getErrors();
@@ -180,27 +224,44 @@ public class ReleaseHelper {
 		context.getProject().release(context.getRepository().getName(), jar);
 		context.getProject().refresh();
 	
-		File[] files = context.getRepository().get(symbName, '[' + version + ',' + version + ']');
-		if (files.length == 1) {
-			ReleaseUtils.toResource(files[0]).refreshLocal(IResource.DEPTH_ZERO, null);
-			jar = new Jar(files[0]);
+		File file = context.getRepository().get(symbName, '[' + version + ',' + version + ']', Strategy.HIGHEST);
+		Jar releasedJar = null;
+		if (file != null) {
+			ReleaseUtils.toResource(file).refreshLocal(IResource.DEPTH_ZERO, null);
+			releasedJar = new Jar(file);
 		}
+		if (releasedJar == null) {
+			handleReleaseErrors(context, context.getProject(), symbName, version);
+			
+			postRelease(context, participants, false);
+			displayErrors(context, Scope.POST_JAR_RELEASE);
+			return false;
+		}
+		context.addReleasedJar(releasedJar);
 		
-		context.addReleasedJar(jar);
-		
-		postJarRelease(context, participants, jar);
+		postJarRelease(context, participants, releasedJar);
 		return true;
 	}
 	
-	private static void updatePackageInfoFile(Project project, PackageInfo packageInfo) throws FileNotFoundException {
+	private static void updatePackageInfoFile(Project project, PackageInfo packageInfo) throws Exception {
 		String path = packageInfo.getPackageName().replace('.', '/') + "/packageinfo";
 		File file = getSourceFile(project, path);
+		
+		// If package/classes are copied into the bundle through Private-Package etc, there will be no source
+		if (!file.getParentFile().exists()) {
+			return;
+		}
+		
 		FileOutputStream fos = new FileOutputStream(file);
 		PrintWriter pw = new PrintWriter(fos);
 		pw.println("version " + packageInfo.getSuggestedVersion());
 		pw.flush();
 		pw.close();
+		ReleaseUtils.toResource(file).refreshLocal(IResource.DEPTH_ZERO, null);
 		
+		File binary = IO.getFile(project.getOutput(), path);
+		IO.copy(file, binary);
+		ReleaseUtils.toResource(binary).refreshLocal(IResource.DEPTH_ZERO, null);
 	}
 	
 	private static File getSourceFile(Project project, String path) {
